requirements.txt
discord.py==2.4.0
Flask==2.2.3
selenium==4.26.1
webdriver_manager==4.0.2
werkzeug==2.2.2
apscheduler

